refactor: storeを作ってそこに処理を逃がす
